+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/subjects/book/BookUiState.kt
@@ -1,10 +1,14 @@
 package com.github.bumblebee202111.doubean.feature.subjects.book
 
-import com.github.bumblebee202111.doubean.model.Book
-import com.github.bumblebee202111.doubean.model.SubjectWithInterest
+import com.github.bumblebee202111.doubean.model.BookDetail
+import com.github.bumblebee202111.doubean.model.SubjectInterestWithUserList
 
 sealed interface BookUiState {
-    data class Success(val book: SubjectWithInterest<Book>, val isLoggedIn: Boolean) : BookUiState
+    data class Success(
+        val book: BookDetail,
+        val interests: SubjectInterestWithUserList,
+        val isLoggedIn: Boolean,
+    ) : BookUiState
     data object Loading : BookUiState
     data object Error : BookUiState
 }