+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/subjects/tv/TvUiState.kt
@@ -1,10 +1,17 @@
 package com.github.bumblebee202111.doubean.feature.subjects.tv
 
-import com.github.bumblebee202111.doubean.model.SubjectWithInterest
-import com.github.bumblebee202111.doubean.model.Tv
+import com.github.bumblebee202111.doubean.model.PhotoList
+import com.github.bumblebee202111.doubean.model.SubjectInterestWithUserList
+import com.github.bumblebee202111.doubean.model.TvDetail
 
 sealed interface TvUiState {
-    data class Success(val tv: SubjectWithInterest<Tv>, val isLoggedIn: Boolean) : TvUiState
+    data class Success(
+        val tv: TvDetail,
+        val interests: SubjectInterestWithUserList,
+        val photos: PhotoList,
+        val isLoggedIn: Boolean,
+    ) : TvUiState
+
     data object Loading : TvUiState
     data object Error : TvUiState
 }